requirements.txt
gradio==3.35.2
openai==0.27.7
python-dotenv==1.0.0
